docs/requirements.txt
audbcards
sphinx
sphinx-audeering-theme >=1.2.1
sphinx-copybutton
